# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Software Api ",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/courses/sessions/{courseSession}/quiz-questions": {
            "get": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Get quiz questions for a specific course session",
                "description": "Retrieve all quiz questions associated with a specific course session, including media and answers.",
                "operationId": "getQuizQuestions",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "ID of the course session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/QuestionResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course session not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/sessions/{courseSession}/save": {
            "post": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Add a course to the user's saved list",
                "description": "Add a course to the authenticated user's saved list.",
                "operationId": "5dfbbf1377c1791d72e812c1c137723b",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "The ID of the course to save",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course successfully added to saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/sessions/{courseSession}/unsave": {
            "delete": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Remove a course from the user's saved list",
                "description": "Remove a course from the authenticated user's saved list.",
                "operationId": "7ebc906b7022a12d220ab87b852ed877",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "The ID of the course to unsave",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Course successfully removed from saved list"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/sessions/{courseSession}/update-progress": {
            "put": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Update user progress for a course session",
                "description": "Update the progress of the authenticated user for a specific course session.",
                "operationId": "updateUserProgress",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "ID of the course session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "complete",
                                    "lastTime"
                                ],
                                "properties": {
                                    "complete": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "lastTime": {
                                        "type": "integer",
                                        "example": 120
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Progress successfully updated"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Course session not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/sessions/{courseSession}": {
            "get": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Get details of a specific course session",
                "description": "Retrieve details of a specific course session, including media and user progress if authenticated.",
                "operationId": "getCourseSession",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "ID of the course session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CourseSessionResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course session not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/courses/sessions/{courseSession}/download/{media}": {
            "post": {
                "tags": [
                    "Course Sessions"
                ],
                "summary": "Download the video associated with a specific course session",
                "description": "Download the video file associated with a specific course session.",
                "operationId": "downloadCourseSessionVideo",
                "parameters": [
                    {
                        "name": "courseSession",
                        "in": "path",
                        "description": "ID of the course session",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "media",
                        "in": "path",
                        "description": "ID of the media file to download",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course session or video not found"
                    }
                }
            }
        },
...
        "/api/courses/my-courses": {
            "get": {
                "tags": [
                    "Courses"
                ],
                "summary": "Get user courses",
                "description": "Returns a list of courses for the authenticated user",
                "operationId": "getUserCourses",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CourseResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CourseResource": {
                "title": "Course Resource",
                "description": "Course resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the course",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "The name of the course",
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "summary": {
                        "description": "A brief summary of the course",
                        "type": "string",
                        "example": "Learn the basics of Laravel framework"
                    },
                    "studentsCount": {
                        "description": "The number of students enrolled in the course",
                        "type": "integer",
                        "example": 25
                    },
                    "createdAt": {
                        "description": "The date and time when the course was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-10-01T12:00:00Z"
                    },
                    "teachers": {
                        "description": "List of teachers associated with the course",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TeacherResource"
                        }
                    },
                    "primaryImage": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "images": {
                        "description": "List of images associated with the course",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "quizezPercent": {
                        "description": "The percentage of quizzes completed by the student",
                        "type": "number",
                        "format": "float",
                        "example": 75.5
                    },
                    "lecturesPercent": {
                        "description": "The percentage of lectures completed by the student",
                        "type": "number",
                        "format": "float",
                        "example": 60
                    },
                    "type": {
                        "description": "The type of the course",
                        "type": "string",
                        "example": "online"
                    },
                    "discount": {
                        "description": "The discount applied to the course",
                        "type": "number",
                        "format": "float",
                        "example": 10
                    },
                    "price": {
                        "description": "The price of the course",
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "rating": {
                        "description": "The rating of the course",
                        "type": "number",
                        "format": "float",
                        "example": 4.5
                    },
                    "teachersCount": {
                        "description": "The number of teachers associated with the course",
                        "type": "integer",
                        "example": 3
                    },
                    "year": {
                        "description": "The year the course was created",
                        "type": "integer",
                        "example": 2023
                    },
                    "time": {
                        "description": "The duration of the course in hours",
                        "type": "integer",
                        "example": 10
                    },
                    "quizezCount": {
                        "description": "The number of quizzes in the course",
                        "type": "integer",
                        "example": 5
                    },
                    "filesCount": {
                        "description": "The number of files associated with the course",
                        "type": "integer",
                        "example": 10
                    }
                },
                "type": "object"
            },
            "CourseSessionResource": {
                "title": "Course Session Resource",
                "description": "Course Session resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the course session",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the course session",
                        "type": "string"
                    },
                    "note": {
                        "description": "Additional notes for the course session",
                        "type": "string"
                    },
                    "time": {
                        "description": "The time of the course session",
                        "type": "string"
                    },
                    "like": {
                        "description": "The number of likes for the course session",
                        "type": "integer"
                    },
                    "dislike": {
                        "description": "The number of dislikes for the course session",
                        "type": "integer"
                    },
                    "type": {
                        "description": "The type of the course session",
                        "type": "string"
                    },
                    "isFree": {
                        "description": "Indicates if the course session is free",
                        "type": "boolean"
                    },
                    "teacher": {
                        "$ref": "#/components/schemas/TeacherResource"
                    },
                    "course": {
                        "$ref": "#/components/schemas/CourseResource"
                    },
                    "video": {
                        "description": "The video media associated with the course session",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MediaResource"
                        }
                    },
                    "image": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "currentUserProgress": {
                        "$ref": "#/components/schemas/UserProgressResource"
                    }
                },
                "type": "object"
            },
            "MediaResource": {
                "title": "MediaResource",
                "description": "Media resource representation for files and images",
                "required": [
                    "id",
                    "name",
                    "fileName",
                    "collection",
                    "url",
                    "size",
                    "type",
                    "extension",
                    "caption"
                ],
                "properties": {
                    "id": {
                        "description": "Unique identifier for the media",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Name of the media file without extension",
                        "type": "string",
                        "example": "nature_photo"
                    },
                    "fileName": {
                        "description": "Complete file name with extension",
                        "type": "string",
                        "example": "nature_photo.jpg"
                    },
                    "collection": {
                        "description": "Collection name where the media is stored",
                        "type": "string",
                        "example": "photos"
                    },
                    "url": {
                        "description": "Full URL to access the media file",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/nature_photo.jpg"
                    },
                    "size": {
                        "description": "Human-readable file size",
                        "type": "string",
                        "example": "2.5 MB"
                    },
                    "type": {
                        "description": "Type of media",
                        "type": "string",
                        "enum": [
                            "image",
                            "video",
                            "document",
                            "audio",
                            "other"
                        ],
                        "example": "image"
                    },
                    "extension": {
                        "description": "File extension",
                        "type": "string",
                        "example": "jpg"
                    },
                    "caption": {
                        "description": "Caption or description of the media. Falls back to name if not set",
                        "type": "string",
                        "example": "Beautiful nature photograph"
                    },
                    "width": {
                        "description": "Width of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1920,
                        "nullable": true
                    },
                    "height": {
                        "description": "Height of the image in pixels (only for images)",
                        "type": "integer",
                        "format": "int32",
                        "example": 1080,
                        "nullable": true
                    },
                    "thumbnailUrl": {
                        "description": "URL of the thumbnail version. Only present if 'thumb' conversion exists",
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/storage/media/conversions/nature_photo-thumb.jpg",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "QuestionResource": {
                "title": "Question Resource",
                "description": "Question resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the question",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the question",
                        "type": "string"
                    },
                    "type": {
                        "description": "The type of the question",
                        "type": "string"
                    },
                    "media": {
                        "$ref": "#/components/schemas/MediaResource"
                    },
                    "answers": {
                        "description": "The answers associated with the question",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnswerResource"
                        }
                    },
                    "note": {
                        "description": "Additional notes or comments for the question",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}